requirements.txt
flask==2.3.3
flask-wtf==1.1.1
wtforms==3.0.1
email-validator==2.1.0
werkzeug==2.3.7
langchain>=0.2.0
langchain-google-genai
langchain-huggingface
langchain-community
langchain-text-splitters
google-generativeai>=0.4.1
sentence-transformers>=2.7.0
faiss-cpu==1.9.0.post1
PyPDF2==3.0.1
python-docx==0.8.11
docx2txt==0.8
pandas
openpyxl==3.1.2
numpy
python-dotenv==1.0.0
